add dto and validaton to delete family, update family, and findall families

+++ src/main/java/com/rentalhive/services/family/FamilyServiceImpl.java
@@ -1,5 +1,6 @@
 package com.rentalhive.services.family;
 
+import com.rentalhive.handlers.exceptionHandler.ResourceNotFoundException;
 import com.rentalhive.models.dto.FamilyDTO;
 import com.rentalhive.models.entities.Family;
 import com.rentalhive.repositories.FamilyRepository;
@@ -8,6 +9,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import javax.persistence.EntityExistsException;
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
@@ -19,7 +21,7 @@ public class FamilyServiceImpl implements FamilyService{
     @Override
     public FamilyDTO createFamily(FamilyDTO familyDTO) {
         Optional<Family> foundFamily = Optional.ofNullable(familyRepository.findByName(familyDTO.name()));
-        if(!foundFamily.isPresent()){
+        if(foundFamily.isPresent()){
             throw new EntityExistsException("Family already exists");
         }
         Family family = familyRepository.save(familyDTO.toFamily());
@@ -27,25 +29,43 @@ public class FamilyServiceImpl implements FamilyService{
     }
 
     @Override
-    public Family updateFamily(Long familyId, Family newFamily) {
-        Family family = familyRepository.findById(familyId).get();
-        family.setName(newFamily.getName());
-        return familyRepository.save(family);
+    public FamilyDTO updateFamily(Long id, FamilyDTO newFamilyDTO) {
+        Optional<Family> foundFamily = familyRepository.findById(id);
+        if(foundFamily.isEmpty()){
+            throw new ResourceNotFoundException("Family with id " + id + " doesn't exist");
+        }
+
+        Optional<Family> foundFamilyByName = Optional.ofNullable(familyRepository.findByName(newFamilyDTO.name()));
+        if(foundFamilyByName.isPresent()){
+            throw new EntityExistsException("Family with this name already exists");
+        }
+
+        foundFamily.get().setName(newFamilyDTO.name());
+
+        Family savedFamily = familyRepository.save(foundFamily.get());
+        return FamilyDTO.toFamilyDTO(savedFamily);
     }
 
     @Override
     public void deleteFamily(Long id) {
+        Optional<Family> foundFamily = getFamilyById(id);
+        if (foundFamily.isEmpty()){
+            throw new ResourceNotFoundException("Family with id " + id + " doesn't exist");
+        }
         familyRepository.deleteById(id);
     }
 
     @Override
-    public Family getFamilyById(Long id) {
-        return familyRepository.findById(id).get();
+    public Optional<Family> getFamilyById(Long id) {
+        Optional<Family> foundFamily = familyRepository.findById(id);
+        return foundFamily;
     }
 
     @Override
-    public List<Family> gelAllFamilies() {
-        return familyRepository.findAll();
+    public List<FamilyDTO> gelAllFamilies() {
+        List<FamilyDTO> families = new ArrayList<>();
+        familyRepository.findAll().forEach(family -> families.add(FamilyDTO.toFamilyDTO(family)));
+        return families;
     }
 
     @Override
